Dev (#28)
* Have boot time a Bot attribute

* Allow to reload all cogs at once

* Move error handling and post-invoke to own coroutines

* New presence activities

* Remove whitespace

* Cleanup

* Use BadArgument exception instead

* Add Tic-Tac-Toe minigame

* Add new activities

* Disable debugging play-against-self

* Silence @everyone and @here responses

* Forgot to specify error

* Add HatBot's favorite movie to Watching list

* Fix bug to steal twice a day

* discord.py version exists

* Add way to nicely format timedeltas

* Make DM behavior clearer

* Add custom help command in Cog

* Add Help to startup extentions

* Handle (possibly) corrupted pickle file

* Run figure editing in async mode (close #21)

* Made hatbot less sassy

* Feesh commands rework

* Fix grammar mistake

* Only send time remaining if the member reacts

* New Feesh Steal Logic (#26)

* Bump pillow from 6.1.0 to 6.2.0

Bumps [pillow](https://github.com/python-pillow/Pillow) from 6.1.0 to 6.2.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/6.1.0...6.2.0)


* feesh: reworked steal, uses tanh, standardised on feesh top, higher odds/lower drop at higher diff.

* feesh: changed tabs->4 spaces and top->top_amount

* Reuse already computed value

* Do not override BasicCog __init__

* Revert feesh steal logic

* Initial Christmas commit

* Fixed feesh steal logic (#27)

* Bump pillow from 6.1.0 to 6.2.0

Bumps [pillow](https://github.com/python-pillow/Pillow) from 6.1.0 to 6.2.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/6.1.0...6.2.0)


* feesh: reworked steal, uses tanh, standardised on feesh top, higher odds/lower drop at higher diff.

* feesh: changed tabs->4 spaces and top->top_amount

* Reuse already computed value

* Fix major bug in new feesh steal logic

* Fix random target potentially being None

* Improve commands permissions

* Add csv file example

* Give more accurate name

* Remove unused function

* Save given keys on reload

* Also delete the command

* Define the giveaway time at top level

* Warn of lack of keys in task

* Add docstrings

* Shorten line

* Handle empty member list

* Add giveaway commands

* Add after_loop

* One giveaway per hour

* Give more permissions to members

* More comments

* Print number of remaining games

* Add Christmas to startup extensions

* Fix getting role from member instead of guild

* Use BasicCog instead

* Load and save keys more reliably

* Remove check on group

* Add reload command for game list

* Split remaining games list and count commands

* Add hugs alias

* Add feesh emoji alias!

* Use Duration converter

* Add owner attribute to MyBot

* Return roles from top of list to bottom

* New avatar command

* Only wait for reactions on the confirmation message

* Remove non-working link

* Remove Christmas from startup extensions

+++ cogs/Admin/admin.py
@@ -20,8 +20,6 @@ class Admin(BasicCog):
 
     def __init__(self, bot):
         super().__init__(bot)
-        # self.bot = bot
-        self.boottime = datetime.now()
 
     async def cog_check(self, ctx):
         return await self.bot.is_owner(ctx.author)
@@ -53,13 +51,23 @@ class Admin(BasicCog):
         await ctx.message.add_reaction('\U00002935')  # :arrow_heading_down:
         await self._cogs_manage(ctx, self.bot.unload_extension, module)
 
-    @cogs.command(name='reload')
+    @cogs.group(name='reload', invoke_without_command=True)
     async def cogs_reload(self, ctx, module):
         """Reload an extension."""
 
         # :arrows_counterclockwise:
         await ctx.message.add_reaction('\U0001F504')
         await self._cogs_manage(ctx, self.bot.reload_extension, module)
+
+    @cogs_reload.command(name='all')
+    async def cogs_reload_all(self, ctx):
+        """Reload all currently loaded extensions."""
+
+        # :arrows_counterclockwise:
+        await ctx.message.add_reaction('\U0001F504')
+        loaded_extensions = list(self.bot.cogs.keys())
+        for cog in loaded_extensions:
+            await self._cogs_manage(ctx, self.bot.reload_extension, cog)
 
     async def _cogs_manage(self, ctx, method, module):
         """Helper method to load/unload/reload modules.
@@ -70,33 +78,62 @@ class Admin(BasicCog):
         if not module.startswith('cogs.'):
             module = f'cogs.{module}'
 
-        try:
-            method(module)
-        except Exception as e:
-            exc = f'{type(e).__name__}: {e}'
-            print(f'Failed to {ctx.invoked_with} extension {module}\n{exc}')
-            await ctx.message.add_reaction('\N{CROSS MARK}')
-            raise e
-        else:
-            print(f'Successfully {ctx.invoked_with}ed extension {module}')
+        # try it, and if an exception is raised, .error is called
+        method(module)
+
+    @cogs_load.after_invoke
+    @cogs_unload.after_invoke
+    @cogs_reload.after_invoke
+    async def cogs_after_invoke(self, ctx):
+        module = ctx.args[2]
+        if not ctx.command_failed:
+            print(f'Successfully {ctx.invoked_with}ed extension {module}.')
             await ctx.message.add_reaction('\N{WHITE HEAVY CHECK MARK}')
+
+    @cogs_reload_all.after_invoke
+    async def cogs_reload_all_after_invoke(self, ctx):
+        if not ctx.command_failed:
+            print('Successfully reloaded all extensions.')
+            await ctx.message.add_reaction('\N{WHITE HEAVY CHECK MARK}')
+
+    @cogs_load.error
+    @cogs_unload.error
+    @cogs_reload.error
+    async def cogs_error(self, ctx, error):
+        module = ctx.args[2]
+        exc = f'{type(error).__name__}: {error}'
+        print(f'Failed to {ctx.invoked_with} extension {module}.\n{exc}')
+        await ctx.message.add_reaction('\N{CROSS MARK}')
+        raise error
+
+    @cogs_reload_all.error
+    async def cogs_reload_all_error(self, ctx, error):
+        exc = f'{type(error).__name__}: {error}'
+        print('Failed to reload some extension.')
+        await ctx.message.add_reaction('\N{CROSS MARK}')
+        await ctx.send(exc)  # because there is lots of info
+        raise error
 
     @commands.command()
     async def uptime(self, ctx):
         """Display the uptime in days, and the boot time."""
 
         # message = 'I have been online for {}! (Since {:%Y-%m-%d %H:%M:%S})'
-        uptime_ = datetime.now() - self.boottime
+        uptime_ = datetime.now() - self.bot.boot_time
         out_str = (f'I have been online for {uptime_.days} days! '
-                   f'(Since {self.boottime:%c})')
+                   f'(Since {self.bot.boot_time:%c})')
         await ctx.send(out_str)
 
     @commands.command()
     async def susay(self, ctx, channel: discord.TextChannel, *, message: str):
-        """Send a message in the requested channel as the Bot."""
+        """Send a message in the requested channel as the Bot.
+        If in DMs, will send to any channel with the given ID.
+        If in a guild, will only work in the current one.
+        """
 
         await channel.send(message)
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
 
     @commands.command()
     async def sudo(self, ctx, who: discord.Member, *, command: str):
